try: refactor the test structure
